backend/requirements.txt
# FastAPI
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.9

# ASGI Server
websockets==13.1

# PyTorch
torch==2.7.1
torchvision==0.22.1

# Audio Processing
librosa==0.10.2
pyloudnorm==0.1.1
soundfile==0.12.1

# Video Processing
opencv-python==4.12.0.88
imageio==2.35.1
imageio-ffmpeg==0.5.1
av>=12.0.0
pillow>=10.0.0

# ML/AI
transformers==4.57.3
diffusers==0.34.0
accelerate==1.8.1

# Database
redis==5.2.0
sqlalchemy==2.0.35
asyncpg==0.29.0

# Utilities
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
loguru==0.7.2
psutil>=5.9.0

# HTTP Client
httpx==0.27.0
aiofiles==24.1.0

# LLM/TTS/ASR
openai>=1.0.0
edge-tts>=6.1.0
resampy>=0.4.0

# Task Queue
celery==5.4.0
kombu==5.4.0

# Monitoring
prometheus-client==0.21.0

# Development
pytest==8.3.3
pytest-asyncio==0.24.0
black==24.8.0
isort==5.13.2
einops>=0.8.0
